requirements.txt
websockets
aiohttp
python-dotenv
aiofiles